improvements

+++ kernel/Utils.php
@@ -133,7 +133,7 @@ function store_uploaded_file($table, $field_name, $file) {
 
 function remove_uploaded_file($table, $field_name, $file) {
     $file_url = __DIR__."/files/uploaded/$table/$field_name/$file";
-    unlink($file_url);
+    if(is_file($file_url)) unlink($file_url);
 }
 
 function core_go_to(string $uri) {
